+++ e2e/e2etests/test_stress_btc_deposit.go
@@ -42,7 +42,7 @@ func TestStressBTCDeposit(r *runner.E2ERunner, args []string) {
 		r.Logger.Print("index %d: starting deposit, tx hash: %s", i, txHash.String())
 
 		eg.Go(func() error {
-			return MonitorBTCDeposit(r, txHash, i, time.Now())
+			return monitorBTCDeposit(r, txHash, i, time.Now())
 		})
 	}
 
@@ -54,8 +54,8 @@ func TestStressBTCDeposit(r *runner.E2ERunner, args []string) {
 	r.Logger.Print("all deposits completed")
 }
 
-// MonitorBTCDeposit monitors the deposit of BTC, returns once the deposit is complete
-func MonitorBTCDeposit(r *runner.E2ERunner, hash *chainhash.Hash, index int, startTime time.Time) error {
+// monitorBTCDeposit monitors the deposit of BTC, returns once the deposit is complete
+func monitorBTCDeposit(r *runner.E2ERunner, hash *chainhash.Hash, index int, startTime time.Time) error {
 	cctx := utils.WaitCctxMinedByInTxHash(r.Ctx, hash.String(), r.CctxClient, r.Logger, r.ReceiptTimeout)
 	if cctx.CctxStatus.Status != crosschaintypes.CctxStatus_OutboundMined {
 		return fmt.Errorf(